Add express-app.js file and moved content into it.
